Added DATSA_WEIGHT to constants.  Moved sumarization logic to its controller.
